// src/infra/api/model/addServicoDto.ts
/**
 * Generated by orval v8.4.1 🍺
 * Do not edit manually.
 * Minha API
 * Documentação oficial da API
 * OpenAPI spec version: 1.0.0
 */

export interface AddServicoDto {
  /**
   * @maxLength 500
   * @nullable
   */
  titulo?: string | null;
  /** @nullable */
  descricao?: string | null;
  /** @nullable */
  valor?: string | null;
}